게시물 상세보기 페이지 디자인 적용 및 버그수정

+++ src/components/postDetail/Comments.tsx
@@ -5,6 +5,9 @@ import CommentCard from '~/components/postDetail/CommentCard';
 import CommentInput from '~/components/postDetail/CommentInput.tsx';
 import {
   CommentContainer,
+  CommentCount,
+  CommentCountTitle,
+  CommentCountWrapper,
   Text,
 } from '~/components/postDetail/commentStyle.ts';
 import {
@@ -30,10 +33,9 @@ const Comments = ({
   updateComment,
 }: CommentsProps) => {
   const [commentState, setCommentState] = useState('');
-  console.log(userId);
 
   const commentInputState = useMemo(() => {
-    const disabled = userId ? false : true;
+    const disabled = !userId;
     const placeholder = userId ? loginPlaceholder : nonLoginPlaceholder;
 
     return {
@@ -61,8 +63,10 @@ const Comments = ({
   };
 
   const deleteMyComment = async (commentId: string) => {
-    await deleteComment({ commentId });
-    await updateComment();
+    if (confirm('댓글을 정말 삭제하시겠습니까?')) {
+      await deleteComment({ commentId });
+      await updateComment();
+    }
   };
 
   return (
@@ -74,7 +78,11 @@ const Comments = ({
         isDisabled={commentInputState['disabled']}
         placeholder={commentInputState['placeholder']}
       />
-      <Text>댓글 {comments.length}</Text>
+      <CommentCountWrapper>
+        <CommentCountTitle>댓글</CommentCountTitle>
+        <CommentCount> {comments.length}</CommentCount>
+      </CommentCountWrapper>
+
       {comments &&
         comments.map((comment, index) => (
           <CommentCard
